added sanitizing function to all

+++ product/pro_edit_done.php
@@ -1,3 +1,19 @@
+<?php 
+session_start();
+session_regenerate_id(true);
+if(isset($_SESSION['login'])==false)
+{
+	$to_top='../staff_login/staff_login.html';
+	print'ログインをしてください。<br />';
+	print'<input type="button" onclick="location.href=\''.$to_top.'\'" value="ログイン画面へ" style="width:200">';
+	exit();
+}
+else
+{
+	print $_SESSION['staff_name'];
+	print'さんログイン中<br /><br />';
+}
+?>
 <!DOCTYPE html>
 <html>
 <head>
@@ -10,16 +26,14 @@
 
 try
 {
+require_once('../common/common.php');
+$post=sanitize($_POST);
+$pro_code=$post['code'];
+$pro_name=$post['name'];
+$pro_price=$post['price'];
+$pro_gazou_name_old=$post['gazou_name_old'];
+$pro_gazou_name=$post['gazou_name'];
 
-$pro_code=$_POST['code'];
-$pro_name=$_POST['name'];
-$pro_price=$_POST['price'];
-$pro_gazou_name_old=$_POST['gazou_name_old'];
-$pro_gazou_name=$_POST['gazou_name'];
-
-$pro_code=htmlspecialchars($pro_code);
-$pro_name=htmlspecialchars($pro_name);
-$pro_price=htmlspecialchars($pro_price);
 
 $dsn='mysql:dbname=shop;host=localhost';
 $user='root';
@@ -31,19 +45,22 @@ $sql='UPDATE mst_product SET name=?,price=?,gazou=? WHERE code=?';
 $stmt=$dbh->prepare($sql);
 $data[]=$pro_name;
 $data[]=$pro_price;
-$data[]=$pro_code;
 $data[]=$pro_gazou_name;
+$data[]=$pro_code;
 $stmt->execute($data);
 
 $dbh=null;
 
-if($pro_gazou_name_old!='')
+if($pro_gazou_name_old!=$pro_gazou_name)
 {
-	unlink('./pics/'.$pro_gazou_name_old);
+	if($pro_gazou_name_old!='')
+	{
+		unlink('./pics/'.$pro_gazou_name_old);
+	}
 }
 
+} //<-----end of try
 
-}
 catch (Exception $e)
 {
 	print'ただいま障害により大変ご迷惑をお掛けしております。';
